+++ metax/gui/main_gui.py
@@ -3839,11 +3839,12 @@ class MetaXGUI(ui_main_window.Ui_metaX_main,QtStyleTools):
                                                                 
                     self.save_and_show_js_plot(pic, title)
                 else:
+                    plt_theme = self.comboBox_basic_theme.currentText()
                     ax = BarPlot(self.tfa, theme=self.html_theme).plot_intensity_bar_sns(df = df, width=width, height=height,  # noqa: F841
                                                                 title= '', rename_taxa=rename_taxa, 
                                                                 show_legend=show_legend, font_size=font_size,
                                                                 rename_sample=rename_sample, plot_mean = plot_mean,
-                                                                plot_percent = plot_percent, sub_meta = sub_meta)
+                                                                plot_percent = plot_percent, sub_meta = sub_meta, plt_theme = plt_theme)
             
             elif plot_type == 'get_table':
                 self.show_message('Getting table...')
@@ -3885,17 +3886,21 @@ class MetaXGUI(ui_main_window.Ui_metaX_main,QtStyleTools):
                 show_percentages = self.checkBox_basic_heatmap_plot_upset_show_percentage.isChecked()
                 min_subset_size = self.spinBox_basic_heatmap_plot_upset_min_subset.value()
                 max_subset_rank = self.spinBox_basic_heatmap_plot_upset_max_rank.value()
-                BasicPlot(self.tfa).plot_upset(df = df, title_name = table_name, show_label = True,
+                upset_df = BasicPlot(self.tfa).plot_upset(df = df, title_name = table_name, show_label = True,
                                 width=width, height=height, font_size=font_size,
                                 plot_sample = False, sub_meta = sub_meta,
                                 rename_sample = rename_sample, show_percentages = show_percentages,
                                 min_subset_size = min_subset_size, max_subset_rank = max_subset_rank)
+                # update the table_dict
+                self.update_table_dict(table_name = f'upset_selected({table_name})', df = upset_df)
                 
         except (IndexError, AttributeError):
             error_message = traceback.format_exc()
             self.logger.write_log(f'plot_basic_info_sns error: {error_message}', 'e')
             QMessageBox.warning(self.MainWindow, 'Warning', 'The index is out of range! Please check the settings.')
-            
+        except ValueError as e:
+            if "At least two groups are required for the UpSet plot." in str(e):
+                QMessageBox.warning(self.MainWindow, 'Warning', 'At least two groups are required for the UpSet plot!')
         except Exception:
             error_message = traceback.format_exc()
             self.logger.write_log(f'plot_basic_list error: {error_message}', 'e')
@@ -4623,16 +4628,20 @@ class MetaXGUI(ui_main_window.Ui_metaX_main,QtStyleTools):
                 show_percentages = self.checkBox_basic_plot_upset_show_percentage.isChecked()
                 min_subset_size = self.spinBox_basic_plot_upset_min_subset.value()
                 max_subset_rank = self.spinBox_basic_plot_upset_max_rank.value()
-                BasicPlot(self.tfa).plot_upset(df = df, title_name = title_name, show_label = show_label,
+                upset_df = BasicPlot(self.tfa).plot_upset(df = df, title_name = title_name, show_label = show_label,
                                                width=width, height=height, font_size=font_size,
                                                plot_sample = plot_sample, sub_meta = sub_meta,
                                                rename_sample = rename_sample, show_percentages = show_percentages,
                                                min_subset_size = min_subset_size, max_subset_rank = max_subset_rank)
+                self.update_table_dict(f'upset_all({title_name})', upset_df)
+                
         except (IndexError, AttributeError):
             error_message = traceback.format_exc()
             self.logger.write_log(f'plot_basic_info_sns error: {error_message}', 'e')
             QMessageBox.warning(self.MainWindow, 'Warning', 'The index is out of range! Please check the settings.')
-            
+        except ValueError as e:
+            if "At least two groups are required for the UpSet plot." in str(e):
+                QMessageBox.warning(self.MainWindow, 'Warning', 'At least two groups are required for the UpSet plot!')
         except Exception:
             error_message = traceback.format_exc()
             simplified_message = "An unexpected error occurred. Please check the logs for details."